+++ modules/klaviyopsautomation/klaviyo-v3-sdk/Exception/KlaviyoRateLimitException.php
@@ -2,6 +2,10 @@
 
 namespace KlaviyoV3Sdk\Exception;
 
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
 class KlaviyoRateLimitException extends KlaviyoApiException
 {
 }