+++ modules/ps_mbo/ps_mbo.php
@@ -29,17 +29,15 @@ if (file_exists($autoloadPath)) {
     require_once $autoloadPath;
 }
 
-use LanguageCore as Language;
 use PrestaShop\Module\Mbo\Accounts\Provider\AccountsDataProvider;
 use PrestaShop\Module\Mbo\Addons\Subscriber\ModuleManagementEventSubscriber;
 use PrestaShop\Module\Mbo\Api\Security\AdminAuthenticationProvider;
 use PrestaShop\Module\Mbo\Helpers\Config;
-use PrestaShop\Module\Mbo\Security\PermissionCheckerInterface;
 use PrestaShop\PrestaShop\Adapter\SymfonyContainer;
 use PrestaShopBundle\Event\ModuleManagementEvent;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\Dotenv\Dotenv;
-use TabCore as Tab;
+use PrestaShop\Module\Mbo\Helpers\ErrorHelper;
 
 class ps_mbo extends Module
 {
@@ -47,15 +45,12 @@ class ps_mbo extends Module
     use PrestaShop\Module\Mbo\Traits\UseHooks;
     use PrestaShop\Module\Mbo\Traits\HaveShopOnExternalService;
 
-    public const DEFAULT_ENV = '';
-
     /**
      * @var string
      */
-    public const VERSION = '4.5.1';
+    public const VERSION = '4.7.1';
 
     public const CONTROLLERS_WITH_CONNECTION_TOOLBAR = [
-        'AdminPsMboModule',
         'AdminModulesManage',
         'AdminModulesSf',
     ];
@@ -83,11 +78,6 @@ class ps_mbo extends Module
     private $serviceContainer;
 
     /**
-     * @var PermissionCheckerInterface
-     */
-    protected $permissionChecker;
-
-    /**
      * @var string
      */
     public $imgPath;
@@ -103,7 +93,7 @@ class ps_mbo extends Module
     public function __construct()
     {
         $this->name = 'ps_mbo';
-        $this->version = '4.5.1';
+        $this->version = '4.7.1';
         $this->author = 'PrestaShop';
         $this->tab = 'administration';
         $this->module_key = '6cad5414354fbef755c7df4ef1ab74eb';
@@ -119,7 +109,11 @@ class ps_mbo extends Module
         $this->moduleCacheDir = sprintf('%s/var/modules/%s/', rtrim(_PS_ROOT_DIR_, '/'), $this->name);
 
         $this->displayName = $this->trans('PrestaShop Marketplace in your Back Office', [], 'Modules.Mbo.Global');
-        $this->description = $this->trans('Browse the Addons marketplace directly from your back office to better meet your needs.', [], 'Modules.Mbo.Global');
+        $this->description = $this
+            ->trans('Browse the Addons marketplace directly from your back office to better meet your needs.',
+                [],
+                'Modules.Mbo.Global'
+            );
 
         if (self::checkModuleStatus()) {
             $this->bootHooks();
@@ -138,7 +132,7 @@ class ps_mbo extends Module
         try {
             $this->getService('mbo.ps_accounts.installer')->install();
         } catch (Exception $e) {
-            // For now, do nothing
+            ErrorHelper::reportError($e);
         }
 
         $this->installTables();
@@ -171,7 +165,7 @@ class ps_mbo extends Module
         $this->getAdminAuthenticationProvider()->deletePossibleApiUser();
         $this->getAdminAuthenticationProvider()->clearCache();
 
-        $lockFiles = ['registerShop', 'updateShop'];
+        $lockFiles = ['registerShop', 'updateShop', 'createApiUser'];
         foreach ($lockFiles as $lockFile) {
             if (file_exists($this->moduleCacheDir . $lockFile . '.lock')) {
                 unlink($this->moduleCacheDir . $lockFile . '.lock');
@@ -240,14 +234,6 @@ class ps_mbo extends Module
         return true;
     }
 
-    private function enableByShop(int $shopId)
-    {
-        // Force context to all shops
-        Shop::setContext(Shop::CONTEXT_SHOP, $shopId);
-
-        return parent::enable(true);
-    }
-
     /**
      * Disable Module.
      *
@@ -278,16 +264,9 @@ class ps_mbo extends Module
         return $this->handleTabAction('uninstall');
     }
 
-    private function disableByShop(int $shopId)
-    {
-        // Force context to all shops
-        Shop::setContext(Shop::CONTEXT_SHOP, $shopId);
-
-        return parent::disable(true);
-    }
-
     /**
-     * Override of native function to always retrieve Symfony container instead of legacy admin container on legacy context.
+     * Override of native function to always retrieve Symfony container instead of legacy admin
+     * container on legacy context.
      *
      * {@inheritdoc}
      */
@@ -340,7 +319,8 @@ class ps_mbo extends Module
             return false;
         }
 
-        // If active = 1 in the module table, the module must be associated to at least one shop to be considered as active
+        // If active = 1
+        //in the module table, the module must be associated to at least one shop to be considered as active
         $result = Db::getInstance()->getRow('SELECT m.`id_module` as `active`, ms.`id_module` as `shop_active`
         FROM `' . _DB_PREFIX_ . 'module` m
         LEFT JOIN `' . _DB_PREFIX_ . 'module_shop` ms ON m.`id_module` = ms.`id_module`
@@ -402,6 +382,50 @@ class ps_mbo extends Module
         return true;
     }
 
+    public function getAccountsDataProvider(): ?AccountsDataProvider
+    {
+        try {
+            return $this->getService('mbo.accounts.data_provider');
+        } catch (\Exception $e) {
+            ErrorHelper::reportError($e);
+            return null;
+        }
+    }
+
+    public function postponeTabsTranslations(): void
+    {
+        /**it'
+         * There is an issue for translating tabs during installation :
+         * Active modules translations files are loaded during the kernel boot.
+         * So the installing module translations are not known
+         * So, we postpone the tabs translations for the first time the module's code is executed.
+         */
+        $lockFile = $this->moduleCacheDir . 'translate_tabs.lock';
+        if (!file_exists($lockFile)) {
+            if (!is_dir($this->moduleCacheDir)) {
+                mkdir($this->moduleCacheDir, 0777, true);
+            }
+            $f = fopen($lockFile, 'w+');
+            fclose($f);
+        }
+    }
+
+    private function enableByShop(int $shopId)
+    {
+        // Force context to all shops
+        Shop::setContext(Shop::CONTEXT_SHOP, $shopId);
+
+        return parent::enable(true);
+    }
+
+    private function disableByShop(int $shopId)
+    {
+        // Force context to all shops
+        Shop::setContext(Shop::CONTEXT_SHOP, $shopId);
+
+        return parent::disable(true);
+    }
+
     private function uninstallTables(): bool
     {
         $sqlQueries[] = 'DROP TABLE IF EXISTS `' . _DB_PREFIX_ . 'mbo_api_config`';
@@ -415,11 +439,6 @@ class ps_mbo extends Module
         return true;
     }
 
-    public function getAccountsDataProvider(): AccountsDataProvider
-    {
-        return $this->getService('mbo.accounts.data_provider');
-    }
-
     /**
      * @return void
      */
@@ -429,20 +448,10 @@ class ps_mbo extends Module
         $dotenv->loadEnv(__DIR__ . '/.env');
     }
 
-    public function postponeTabsTranslations(): void
+    private function isPsAccountEnabled(): bool
     {
-        /**it'
-         * There is an issue for translating tabs during installation :
-         * Active modules translations files are loaded during the kernel boot. So the installing module translations are not known
-         * So, we postpone the tabs translations for the first time the module's code is executed.
-         */
-        $lockFile = $this->moduleCacheDir . 'translate_tabs.lock';
-        if (!file_exists($lockFile)) {
-            if (!is_dir($this->moduleCacheDir)) {
-                mkdir($this->moduleCacheDir);
-            }
-            $f = fopen($lockFile, 'w+');
-            fclose($f);
-        }
+        $accountsInstaller = $this->get('mbo.ps_accounts.installer');
+
+        return null !== $accountsInstaller && $accountsInstaller->isModuleEnabled();
     }
 }