+++ modules/klaviyopsautomation/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- BEGIN RELEASE NOTES -->
 ### [Unreleased]
 
+### [1.4.3] - 2023-12-13
+
+#### Fixed
+- Pass Viewed Product event value as float.
+- Paginate to get all lists for account.
+- Use unique hash value in autoloader file.
+
 ### [1.4.2] - 2023-10-20
 
 #### Fixed
@@ -235,7 +242,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- END RELEASE NOTES -->
 
 <!-- BEGIN LINKS -->
-[Unreleased]: https://github.com/klaviyo/prestashop_klaviyo/compare/1.4.2...HEAD
+[Unreleased]: https://github.com/klaviyo/prestashop_klaviyo/compare/1.4.3...HEAD
+[1.4.3]: https://github.com/klaviyo/prestashop_klaviyo/compare/1.4.2...1.4.3
 [1.4.2]: https://github.com/klaviyo/prestashop_klaviyo/compare/1.4.1...1.4.2
 [1.4.1]: https://github.com/klaviyo/prestashop_klaviyo/compare/1.4.0...1.4.1
 [1.4.0]: https://github.com/klaviyo/prestashop_klaviyo/compare/1.3.2...1.4.0